Refactor CircleTransform into Utility class and removed excessive Listeners
